ffmpeg检查

+++ src/main/java/cn/anytec/quadrant/hcService/ImgToAreaVideoRunnable.java
@@ -9,6 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import java.io.File;
+import java.util.Arrays;
+import java.util.function.Predicate;
 
 @Component
 public class ImgToAreaVideoRunnable implements Runnable {
@@ -38,7 +40,7 @@ public class ImgToAreaVideoRunnable implements Runnable {
         while (true) {
             try {
                 File[] areaVideoFlies = areaFile.listFiles();
-                if (areaVideoFlies == null) {
+                if (areaVideoFlies == null || areaVideoFlies.length == 0) {
                     Thread.sleep(2000);
                     continue;
                 }
@@ -61,6 +63,11 @@ public class ImgToAreaVideoRunnable implements Runnable {
                                     .append(areaVideoName);
                             logger.info("开始生成" + areaVideoName);
                             logger.debug(runtimeLocal.execute(makeAreaVideo.toString()));
+                            Predicate<String> isExist = (n) -> n.endsWith(".mp4");
+                            if(!Arrays.stream(areaVideos).anyMatch(isExist)){
+                                logger.error("生成视频失败！检查是否安装了ffmpeg并重启应用");
+                                return;
+                            }
                             new File(cameraIpFile,areaVideo).delete();
 
                         }
